[#34718] Fix coupling with code_saturne

+++ code_aster/Coupling/parameters.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 # --------------------------------------------------------------------
-# Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
+# Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 # This file is part of code_aster.
 #
 # code_aster is free software: you can redistribute it and/or modify
@@ -74,7 +74,7 @@ class SchemeParams:
 
                 if nb_step is None:
                     assert final_time is not None
-                    nb_step = (final_time - init_time) / delta_t
+                    nb_step = int((final_time - init_time) / delta_t)
 
                 time_list = np.linspace(init_time, final_time, nb_step + 1, endpoint=True)
